requirements.txt
streamlit==1.34.0
pandas
Pillow
flask
flask-cors
openpyxl